[py] move proxy and insecure certs capabilities to Base Options
